!WIP!
tagging gallery/models/item.php (main item model) for places which use physical path in a url, or assume it will be inside VARPATH . "albums/"
these will need to all change to use the resource_Controller which will handle thumbs, resized images, and the main image as well.  the point of this is to obscure the actual file's location on disk from the browser.  It will be passed as a resource stream.